chore: sync loans polaris script improved
